Move MCPServerViewsContext to shared folder
