+++ front/components/agent_builder/capabilities/usePresetActionHandler.ts
@@ -6,7 +6,7 @@ import type {
   AgentBuilderAction,
   AgentBuilderFormData,
 } from "@app/components/agent_builder/AgentBuilderFormContext";
-import { useMCPServerViewsContext } from "@app/components/agent_builder/MCPServerViewsContext";
+import { useMCPServerViewsContext } from "@app/components/shared/tools_picker/MCPServerViewsContext";
 import { getDefaultMCPAction } from "@app/components/agent_builder/types";
 import { useSendNotification } from "@app/hooks/useNotification";
 import {
